Atividades de métodos

+++ src/curso_programacao/MethodCalculoTriangulo.java
@@ -10,14 +10,14 @@ public class MethodCalculoTriangulo {
 
 		Locale.setDefault(Locale.US);
 
-		int retaA, retaB, retaC;
+		Double retaA, retaB, retaC;
 
 		System.out.println("Tamanho da reta A :");
-		retaA = sc.nextInt();
+		retaA = sc.nextDouble();
 		System.out.println("Tamanho da reta B :");
-		retaB = sc.nextInt();
+		retaB = sc.nextDouble();
 		System.out.println("Tamanho da reta C :");
-		retaC = sc.nextInt();
+		retaC = sc.nextDouble();
 
 		verificarSeTriangulo(retaA, retaB, retaC);
 		verificarLadosTriangulo(retaA, retaB, retaC);
@@ -27,7 +27,7 @@ public class MethodCalculoTriangulo {
 
 	}
 
-	public static void verificarSeTriangulo(int retaA, int retaB, int retaC) {
+	public static void verificarSeTriangulo(Double retaA, Double retaB, Double retaC) {
 		if (retaA < retaB + retaC || retaB < retaC + retaA || retaC < retaA + retaB) {
 			System.out.println("-------------------------");
 			System.out.println("Trata-se de um TRIÂNGULO.");
@@ -36,7 +36,7 @@ public class MethodCalculoTriangulo {
 		
 	}
 
-	public static void verificarLadosTriangulo(int retaA, int retaB, int retaC) {
+	public static void verificarLadosTriangulo(Double retaA, Double retaB, Double retaC) {
 		if (retaA == retaB && retaA == retaC && retaB == retaC) {
 			System.out.println("O triângulo tem 3 lados iguais, portanto ele é um triangulo EQUILÁTERO.");
 		} else if (retaA == retaB || retaA == retaC || retaB == retaC) {
